Split tools.py (step 2)
- _roi.py lower level file, tools for dealing with ROIs
- _overlap.py higher level file, computing overlaps for reprojection
